+++ src/core/promise/abortable/spec.js
@@ -6,14 +6,14 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
-import Then from 'core/promise/abortable';
+import AbortablePromise from 'core/promise/abortable';
 
 describe('core/promise/abortable', () => {
-	it('simple then', async () => {
+	it('simple `then`', async () => {
 		let
 			i = 1;
 
-		const promise = new Then((resolve) => {
+		const promise = new AbortablePromise((resolve) => {
 			resolve();
 		})
 			.then(() => i + 2)
@@ -25,11 +25,58 @@ describe('core/promise/abortable', () => {
 		expect(i).toBe(6);
 	});
 
+	it('promise that is resolved with another promise', async () => {
+		const i = await new AbortablePromise((resolve) => {
+			resolve(
+				new Promise((r) => setTimeout(() => r(AbortablePromise.resolve(1)), 50))
+			);
+		})
+			.then((val) => new Promise((r) => setTimeout(() => r(AbortablePromise.resolve(val + 2)), 50)))
+			.then((val) => val * 2);
+
+		expect(i).toBe(6);
+	});
+
+	it('promise that is rejected with another promise', async () => {
+		let i;
+
+		try {
+			await new AbortablePromise((resolve) => {
+				resolve(
+					new Promise((r) => setTimeout(() => r(AbortablePromise.resolve(1)), 50))
+				);
+			})
+				.then((val) => new Promise((r) => setTimeout(() => r(AbortablePromise.reject(val + 2)), 50)))
+				.catch((val) => Promise.reject(val * 2));
+
+		} catch (err) {
+			i = err;
+		}
+
+		expect(i).toBe(6);
+	});
+
+	it('promise that is rejected with another promise by using a constructor', async () => {
+		let i;
+
+		try {
+			await new AbortablePromise((resolve, reject) => {
+				reject(AbortablePromise.resolve(1));
+			});
+
+		} catch (err) {
+			i = err;
+		}
+
+		expect(i).toBeInstanceOf(AbortablePromise);
+		expect(await i).toBe(1);
+	});
+
 	it('aborting of a promise', async () => {
 		let
 			status = 'pending';
 
-		const promise = new Then((resolve, reject, onAbort) => {
+		const promise = new AbortablePromise((resolve, reject, onAbort) => {
 			onAbort(() => {
 				status = 'aborted';
 			});
@@ -53,8 +100,8 @@ describe('core/promise/abortable', () => {
 	it('providing of a parent promise', async () => {
 		try {
 			const
-				parentPromise = new Then(() => undefined),
-				promise = new Then(() => undefined, parentPromise).catch((err) => err);
+				parentPromise = new AbortablePromise(() => undefined),
+				promise = new AbortablePromise(() => undefined, parentPromise).catch((err) => err);
 
 			parentPromise.abort('boom');
 			expect(await promise)
@@ -66,8 +113,8 @@ describe('core/promise/abortable', () => {
 
 		try {
 			const
-				parentPromise = new Then(() => undefined).catch((err) => err),
-				promise = new Then(() => undefined, parentPromise);
+				parentPromise = new AbortablePromise(() => undefined).catch((err) => err),
+				promise = new AbortablePromise(() => undefined, parentPromise);
 
 			promise.abort('boom');
 			expect(await parentPromise).toBe('boom');
@@ -80,7 +127,7 @@ describe('core/promise/abortable', () => {
 		let
 			i = 1;
 
-		await new Then((resolve, reject) => {
+		await new AbortablePromise((resolve, reject) => {
 			reject('boom');
 		})
 			.then((val) => val * 2, (err) => {
@@ -95,7 +142,7 @@ describe('core/promise/abortable', () => {
 		let
 			i = 1;
 
-		await new Then((resolve) => {
+		await new AbortablePromise((resolve) => {
 			resolve();
 		})
 			.then(() => {
@@ -113,11 +160,11 @@ describe('core/promise/abortable', () => {
 		expect(i).toBe(6);
 	});
 
-	it('catch', async () => {
+	it('`catch`', async () => {
 		let
 			i = 1;
 
-		await new Then((resolve, reject) => {
+		await new AbortablePromise((resolve, reject) => {
 			reject('boom');
 		})
 			.then((val) => val * 2)
@@ -133,7 +180,7 @@ describe('core/promise/abortable', () => {
 		let
 			i = 1;
 
-		await new Then((resolve) => {
+		await new AbortablePromise((resolve) => {
 			resolve();
 		})
 			.then(() => {
@@ -155,7 +202,7 @@ describe('core/promise/abortable', () => {
 		let
 			i = 1;
 
-		await new Then((resolve) => {
+		await new AbortablePromise((resolve) => {
 			resolve();
 		})
 			.then(() => i + 2)
@@ -172,7 +219,7 @@ describe('core/promise/abortable', () => {
 			i = 1;
 
 		try {
-			await new Then((resolve, reject) => {
+			await new AbortablePromise((resolve, reject) => {
 				reject('boom');
 			})
 				.finally((arg) => {
@@ -191,7 +238,7 @@ describe('core/promise/abortable', () => {
 		let
 			i = 1;
 
-		await new Then((resolve) => {
+		await new AbortablePromise((resolve) => {
 			resolve();
 		})
 			.then(() => {
@@ -218,16 +265,16 @@ describe('core/promise/abortable', () => {
 		expect(i).toBe(6);
 	});
 
-	it('`Then.resolve`', async () => {
+	it('`AbortablePromise.resolve`', async () => {
 		let
 			i = 0,
 			j = 0;
 
-		await Then.resolve(1)
+		await AbortablePromise.resolve(1)
 			.then((val) => i = val + 2)
 			.then((val) => i = val * 2);
 
-		await Then.resolve(Then.resolve(1))
+		await AbortablePromise.resolve(AbortablePromise.resolve(1))
 			.then((val) => j = val + 2)
 			.then((val) => j = val * 2);
 
@@ -235,11 +282,11 @@ describe('core/promise/abortable', () => {
 		expect(j).toBe(6);
 	});
 
-	it('`Then.reject`', async () => {
+	it('`AbortablePromise.reject`', async () => {
 		let
 			i = 1;
 
-		await Then.reject('boom')
+		await AbortablePromise.reject('boom')
 			.catch((err) => {
 				expect(err).toBe('boom');
 				i += 2;
@@ -248,12 +295,12 @@ describe('core/promise/abortable', () => {
 		expect(i).toBe(3);
 	});
 
-	describe('`Then.all`', () => {
+	describe('`AbortablePromise.all`', () => {
 		it('all promises are resolved', async () => {
-			const res = await Then.all([
+			const res = await AbortablePromise.all([
 				1,
 				null,
-				Then.resolve(2)
+				AbortablePromise.resolve(2)
 			]);
 
 			expect(res).toEqual([1, null, 2]);
@@ -264,10 +311,10 @@ describe('core/promise/abortable', () => {
 				res;
 
 			try {
-				res = await Then.all([
+				res = await AbortablePromise.all([
 					1,
 					null,
-					Then.reject(2)
+					AbortablePromise.reject(2)
 				]);
 
 			} catch (err) {
@@ -278,22 +325,22 @@ describe('core/promise/abortable', () => {
 		});
 	});
 
-	it('`Then.race`', async () => {
+	it('`AbortablePromise.race`', async () => {
 		let res;
 
-		await Then.race([
+		await AbortablePromise.race([
 			Promise.resolve(1),
-			Then.resolve(2)
+			AbortablePromise.resolve(2)
 		]).then((val) => res = val);
 
 		expect(res).toBe(2);
 	});
 
-	describe('`Then.race`', () => {
+	describe('`AbortablePromise.race`', () => {
 		it('all promises are resolved', async () => {
-			const res = await Then.race([
+			const res = await AbortablePromise.race([
 				Promise.resolve(1),
-				Then.resolve(2)
+				AbortablePromise.resolve(2)
 			]);
 
 			expect(res).toBe(2);
@@ -304,9 +351,9 @@ describe('core/promise/abortable', () => {
 				res;
 
 			try {
-				res = await Then.race([
+				res = await AbortablePromise.race([
 					new Promise((r) => setTimeout(r, 15)),
-					Then.reject(2)
+					AbortablePromise.reject(2)
 				]);
 
 			} catch (err) {