+++ src/core/promise/sync/CHANGELOG.md
@@ -15,6 +15,10 @@ Changelog
 
 * Added a new static method `any`
 
+#### :bug: Bug Fix
+
+* Fixed some cases when a promise is resolved with another promise
+
 ## v3.58.0 (2021-09-01)
 
 #### :rocket: New Feature